refactor : pr적용
