+++ src/docs/asciidoc/api/user/user-nickname-change.adoc
@@ -1,4 +1,3 @@
-=== 유저 API ===
 
 ==== 닉네임 변경 ====
 
@@ -11,13 +10,6 @@
 - 닉네임은 한글, 영문, 숫자만 사용할 수 있습니다.
 - 닉네임은 공백을 포함할 수 없습니다.
 - userId와 newNickname은 필수 요청 필드입니다.
-
-
-|====
-|속성        |설명
-|`userId`    |사용자 식별자
-|`newNickname`|새로운 닉네임
-|====
 
 [source]
 ----
