+++ src/main/java/app/bottlenote/user/dto/response/NicknameChangeResponse.java
@@ -8,25 +8,15 @@ import lombok.Getter;
 public class NicknameChangeResponse {
 
 	private final String message;
-	@JsonProperty("userId")
+
 	private Long userId;
 
-	@JsonProperty("beforeNickname")
 	private String beforeNickname;
 
-	@JsonProperty("changedNickname")
 	private String changedNickname;
 
-	protected NicknameChangeResponse(String message, Long userId, String beforeNickname, String changedNickname) {
-		this.message = message;
-		this.userId = userId;
-		this.beforeNickname = beforeNickname;
-		this.changedNickname = changedNickname;
-	}
-
-
 	public static NicknameChangeResponse of(
-		NicknameChangeResponseEnum message
+		Message message
 		, Long userId
 		, String beforeNickname
 		, String changedNickname) {
@@ -35,7 +25,9 @@ public class NicknameChangeResponse {
 
 	@AllArgsConstructor
 	@Getter
-	public enum NicknameChangeResponseEnum {
+	// NicknameChangeResponseEnum
+	// NicknameChangeMessage
+	public enum Message {
 		SUCCESS("닉네임이 성공적으로 변경되었습니다."),
 		FAIL("닉네임 변경에 실패하였습니다.");
 		private final String message;
